+++ src/libtropic.c
@@ -77,20 +77,30 @@ lt_ret_t lt_update_mode(lt_handle_t *h)
         return LT_PARAM_ERR;
     }
 
+    lt_ret_t ret;
+
     // The byte used here must not be ID byte of some request, otherwise chip would be confused
     // and would return CRC error.
     // GET_RESP 0xAA works fine.
     h->l2.buff[0] = GET_RESPONSE_REQ_ID;
 
     // Transfer just one byte to read CHIP_STATUS byte
-    lt_l1_spi_csn_low(&h->l2);
-
-    if (lt_l1_spi_transfer(&h->l2, 0, 1, LT_L1_TIMEOUT_MS_DEFAULT) != LT_OK) {
-        lt_l1_spi_csn_high(&h->l2);
-        return LT_L1_SPI_ERROR;
+    ret = lt_l1_spi_csn_low(&h->l2);
+    if (ret != LT_OK) {
+        return ret;
     }
 
-    lt_l1_spi_csn_high(&h->l2);
+    ret = lt_l1_spi_transfer(&h->l2, 0, 1, LT_L1_TIMEOUT_MS_DEFAULT);
+    if (ret != LT_OK) {
+        lt_ret_t ret_unused = lt_l1_spi_csn_high(&h->l2);
+        UNUSED(ret_unused);  // We don't care about it, we return ret from SPI transfer anyway.
+        return ret;
+    }
+
+    ret = lt_l1_spi_csn_high(&h->l2);
+    if (ret != LT_OK) {
+        return ret;
+    }
 
     // Buffer in handle now contains CHIP_STATUS byte,
     // Save info about chip mode into 'mode' variable in handle
@@ -253,9 +263,9 @@ lt_ret_t lt_get_info_chip_id(lt_handle_t *h, struct lt_chip_id_t *chip_id)
     return LT_OK;
 }
 
-lt_ret_t lt_get_info_riscv_fw_ver(lt_handle_t *h, uint8_t *ver, const uint16_t max_len)
+lt_ret_t lt_get_info_riscv_fw_ver(lt_handle_t *h, uint8_t *ver)
 {
-    if (!h || !ver || max_len < LT_L2_GET_INFO_RISCV_FW_SIZE) {
+    if (!h || !ver) {
         return LT_PARAM_ERR;
     }
 
@@ -288,9 +298,9 @@ lt_ret_t lt_get_info_riscv_fw_ver(lt_handle_t *h, uint8_t *ver, const uint16_t m
     return LT_OK;
 }
 
-lt_ret_t lt_get_info_spect_fw_ver(lt_handle_t *h, uint8_t *ver, const uint16_t max_len)
+lt_ret_t lt_get_info_spect_fw_ver(lt_handle_t *h, uint8_t *ver)
 {
-    if (!h || !ver || max_len < LT_L2_GET_INFO_SPECT_FW_SIZE) {
+    if (!h || !ver) {
         return LT_PARAM_ERR;
     }
 
@@ -489,7 +499,10 @@ lt_ret_t lt_reboot(lt_handle_t *h, const uint8_t startup_id)
         return LT_FAIL;
     }
 
-    lt_l1_delay(&h->l2, LT_TROPIC01_REBOOT_DELAY_MS);
+    ret = lt_l1_delay(&h->l2, LT_TROPIC01_REBOOT_DELAY_MS);
+    if (ret != LT_OK) {
+        return ret;
+    }
 
     // Update mode variable in handle after reboot
     ret = lt_update_mode(h);
@@ -668,9 +681,9 @@ lt_ret_t lt_mutable_fw_update_data(lt_handle_t *h, const uint8_t *update_data, c
 #error "Undefined silicon revision. Please define either ABAB or ACAB."
 #endif
 
-lt_ret_t lt_get_log(lt_handle_t *h, uint8_t *log_msg, uint16_t msg_len_max)
+lt_ret_t lt_get_log_req(lt_handle_t *h, uint8_t *log_msg, uint16_t *log_msg_len)
 {
-    if (!h || !log_msg || msg_len_max > GET_LOG_MAX_MSG_LEN) {
+    if (!h || !log_msg || !log_msg_len) {
         return LT_PARAM_ERR;
     }
 
@@ -691,9 +704,7 @@ lt_ret_t lt_get_log(lt_handle_t *h, uint8_t *log_msg, uint16_t msg_len_max)
         return ret;
     }
 
-    // No check for incomming l3 length because we don't know in advance how big message will be,
-    // the max possible length is 255 (uint8_t) and that fits the safe size GET_LOG_MAX_MSG_LEN of log_msg buffer
-
+    *log_msg_len = p_l2_resp->rsp_len;
     memcpy(log_msg, p_l2_resp->log_msg, p_l2_resp->rsp_len);
 
     return LT_OK;
@@ -1375,6 +1386,7 @@ static const char *lt_ret_strs[] = {"LT_OK",
                                     "LT_L2_IN_CRC_ERR",
                                     "LT_L2_REQ_CONT",
                                     "LT_L2_RES_CONT",
+                                    "LT_L2_RESP_DISABLED",
                                     "LT_L2_HSK_ERR",
                                     "LT_L2_NO_SESSION",
                                     "LT_L2_TAG_ERR",
@@ -1384,7 +1396,7 @@ static const char *lt_ret_strs[] = {"LT_OK",
                                     "LT_L2_UNKNOWN_REQ",
                                     "LT_L2_STATUS_NOT_RECOGNIZED",
                                     "LT_L2_DATA_LEN_ERROR",
-                                    "LT_CERT_STORE_INVALID"
+                                    "LT_CERT_STORE_INVALID",
                                     "LT_CERT_UNSUPPORTED",
                                     "LT_CERT_ITEM_NOT_FOUND",
                                     "LT_NONCE_OVERFLOW"};
@@ -1526,14 +1538,14 @@ lt_ret_t lt_verify_chip_and_start_secure_session(lt_handle_t *h, uint8_t *shipri
 
     // This is not used in this example, but let's read it anyway
     uint8_t riscv_fw_ver[LT_L2_GET_INFO_RISCV_FW_SIZE] = {0};
-    ret = lt_get_info_riscv_fw_ver(h, riscv_fw_ver, LT_L2_GET_INFO_RISCV_FW_SIZE);
+    ret = lt_get_info_riscv_fw_ver(h, riscv_fw_ver);
     if (ret != LT_OK) {
         return ret;
     }
 
     // This is not used in this example, but let's read it anyway
     uint8_t spect_fw_ver[LT_L2_GET_INFO_SPECT_FW_SIZE] = {0};
-    ret = lt_get_info_spect_fw_ver(h, spect_fw_ver, LT_L2_GET_INFO_SPECT_FW_SIZE);
+    ret = lt_get_info_spect_fw_ver(h, spect_fw_ver);
     if (ret != LT_OK) {
         return ret;
     }
@@ -1598,7 +1610,7 @@ lt_ret_t lt_print_chip_id(const struct lt_chip_id_t *chip_id, int (*print_func)(
 
     if (LT_OK
             != lt_print_bytes(chip_id->chip_id_ver, sizeof(chip_id->chip_id_ver), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("CHIP_ID ver            = 0x%s (v%" PRIu8 ".%" PRIu8 ".%" PRIu8 ".%" PRIu8 ")\r\n",
                           print_bytes_buff, chip_id->chip_id_ver[0], chip_id->chip_id_ver[1], chip_id->chip_id_ver[2],
                           chip_id->chip_id_ver[3])) {
@@ -1607,7 +1619,7 @@ lt_ret_t lt_print_chip_id(const struct lt_chip_id_t *chip_id, int (*print_func)(
 
     if (LT_OK
             != lt_print_bytes(chip_id->fl_chip_info, sizeof(chip_id->fl_chip_info), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("FL_PROD_DATA           = 0x%s (%s)\r\n", print_bytes_buff,
                           chip_id->fl_chip_info[0] == 0x01 ? "PASSED" : "N/A")) {
         return LT_FAIL;
@@ -1615,7 +1627,7 @@ lt_ret_t lt_print_chip_id(const struct lt_chip_id_t *chip_id, int (*print_func)(
 
     if (LT_OK
             != lt_print_bytes(chip_id->func_test_info, sizeof(chip_id->func_test_info), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("MAN_FUNC_TEST          = 0x%s (%s)\r\n", print_bytes_buff,
                           chip_id->func_test_info[0] == 0x01 ? "PASSED" : "N/A")) {
         return LT_FAIL;
@@ -1623,7 +1635,7 @@ lt_ret_t lt_print_chip_id(const struct lt_chip_id_t *chip_id, int (*print_func)(
 
     if (LT_OK
             != lt_print_bytes(chip_id->silicon_rev, sizeof(chip_id->silicon_rev), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("Silicon rev            = 0x%s (%c%c%c%c)\r\n", print_bytes_buff, chip_id->silicon_rev[0],
                           chip_id->silicon_rev[1], chip_id->silicon_rev[2], chip_id->silicon_rev[3])) {
         return LT_FAIL;
@@ -1632,7 +1644,7 @@ lt_ret_t lt_print_chip_id(const struct lt_chip_id_t *chip_id, int (*print_func)(
     uint16_t packg_type_id = ((uint16_t)chip_id->packg_type_id[0] << 8) | ((uint16_t)chip_id->packg_type_id[1]);
     if (LT_OK
         != lt_print_bytes(chip_id->packg_type_id, sizeof(chip_id->packg_type_id), print_bytes_buff,
-                          CHIP_ID_FIELD_MAX_SIZE)) {
+                          sizeof(print_bytes_buff))) {
         return LT_FAIL;
     }
     char packg_type_id_str[17];
@@ -1688,24 +1700,25 @@ lt_ret_t lt_print_chip_id(const struct lt_chip_id_t *chip_id, int (*print_func)(
 
     if (LT_OK
             != lt_print_bytes(chip_id->provisioning_date, sizeof(chip_id->provisioning_date), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("Prov date              = 0x%s \r\n", print_bytes_buff)) {
         return LT_FAIL;
     }
 
-    if (LT_OK != lt_print_bytes(chip_id->hsm_ver, sizeof(chip_id->hsm_ver), print_bytes_buff, CHIP_ID_FIELD_MAX_SIZE)
+    if (LT_OK != lt_print_bytes(chip_id->hsm_ver, sizeof(chip_id->hsm_ver), print_bytes_buff, sizeof(print_bytes_buff))
         || 0 > print_func("HSM HW/FW/SW ver       = 0x%s\r\n", print_bytes_buff)) {
         return LT_FAIL;
     }
 
-    if (LT_OK != lt_print_bytes(chip_id->prog_ver, sizeof(chip_id->prog_ver), print_bytes_buff, CHIP_ID_FIELD_MAX_SIZE)
+    if (LT_OK
+            != lt_print_bytes(chip_id->prog_ver, sizeof(chip_id->prog_ver), print_bytes_buff, sizeof(print_bytes_buff))
         || 0 > print_func("Programmer ver         = 0x%s\r\n", print_bytes_buff)) {
         return LT_FAIL;
     }
 
     if (LT_OK
             != lt_print_bytes((uint8_t *)&chip_id->ser_num, sizeof(chip_id->ser_num), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("S/N                    = 0x%s\r\n", print_bytes_buff)) {
         return LT_FAIL;
     }
@@ -1716,14 +1729,14 @@ lt_ret_t lt_print_chip_id(const struct lt_chip_id_t *chip_id, int (*print_func)(
     pn_data[pn_len] = '\0';
     if (LT_OK
             != lt_print_bytes(chip_id->part_num_data, sizeof(chip_id->part_num_data), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("P/N (long)             = 0x%s (%s)\r\n", print_bytes_buff, pn_data)) {
         return LT_FAIL;
     }
 
     if (LT_OK
             != lt_print_bytes(chip_id->prov_templ_ver, sizeof(chip_id->prov_templ_ver), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("Prov template ver      = 0x%s (v%" PRIu8 ".%" PRIu8 ")\r\n", print_bytes_buff,
                           chip_id->prov_templ_ver[0], chip_id->prov_templ_ver[1])) {
         return LT_FAIL;
@@ -1731,14 +1744,14 @@ lt_ret_t lt_print_chip_id(const struct lt_chip_id_t *chip_id, int (*print_func)(
 
     if (LT_OK
             != lt_print_bytes(chip_id->prov_templ_tag, sizeof(chip_id->prov_templ_tag), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("Prov template tag      = 0x%s\r\n", print_bytes_buff)) {
         return LT_FAIL;
     }
 
     if (LT_OK
             != lt_print_bytes(chip_id->prov_spec_ver, sizeof(chip_id->prov_spec_ver), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("Prov specification ver = 0x%s (v%" PRIu8 ".%" PRIu8 ")\r\n", print_bytes_buff,
                           chip_id->prov_spec_ver[0], chip_id->prov_spec_ver[1])) {
         return LT_FAIL;
@@ -1746,12 +1759,13 @@ lt_ret_t lt_print_chip_id(const struct lt_chip_id_t *chip_id, int (*print_func)(
 
     if (LT_OK
             != lt_print_bytes(chip_id->prov_spec_tag, sizeof(chip_id->prov_spec_tag), print_bytes_buff,
-                              CHIP_ID_FIELD_MAX_SIZE)
+                              sizeof(print_bytes_buff))
         || 0 > print_func("Prov specification tag = 0x%s\r\n", print_bytes_buff)) {
         return LT_FAIL;
     }
 
-    if (LT_OK != lt_print_bytes(chip_id->batch_id, sizeof(chip_id->batch_id), print_bytes_buff, CHIP_ID_FIELD_MAX_SIZE)
+    if (LT_OK
+            != lt_print_bytes(chip_id->batch_id, sizeof(chip_id->batch_id), print_bytes_buff, sizeof(print_bytes_buff))
         || 0 > print_func("Batch ID               = 0x%s\r\n", print_bytes_buff)) {
         return LT_FAIL;
     }