+++ src/lt_crc16.h
@@ -18,7 +18,7 @@
  * @param size      Length if data in buffer
  * @return          CRC16 checksum returned as uint16_t
  */
-uint16_t crc16(const uint8_t *buf, int16_t size);
+uint16_t crc16(const uint8_t *buf, int16_t size) __attribute__((warn_unused_result));
 
 /**
  * @brief Takes pointer to filled l2 buffer and adds checksum